requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
grpcio @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_35t8120i87/croot/grpc-split_1716834576536/work
grpcio-tools @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_ac0kdi6u0e/croot/grpcio-tools_1721915941488/work
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
wheel==0.45.1
